+++ extension/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## Unreleased
+- Support for tracking channels per repository/extension.
+
 ## 0.9.0-beta.0
 
 - Initial release